+++ src/lib9/fmt/sprint.c
@@ -36,10 +36,10 @@ sprint(char *buf, char *fmt, ...)
 	 * optimizes the test away.  casting to uintptr works around this bug.
 	 */
 	if((uintptr)buf+len < (uintptr)buf)
-		len = -(uintptr)buf-1;
+		len = (uint)-(uintptr)buf-1;
 
 	va_start(args, fmt);
-	n = vsnprint(buf, len, fmt, args);
+	n = (int)vsnprint(buf, (int)len, fmt, args);
 	va_end(args);
 	return n;
 }
